+++ seed/ts-sdk/error-property/union-utils/src/api/resources/errors/errors/PropertyBasedErrorTest.ts
@@ -5,6 +5,8 @@ import * as errors from "../../../../errors/index.js";
 import type * as SeedErrorProperty from "../../../index.js";
 
 export class PropertyBasedErrorTest extends errors.SeedErrorPropertyError {
+    public declare readonly body: SeedErrorProperty.PropertyBasedErrorTestBody;
+
     constructor(body: SeedErrorProperty.PropertyBasedErrorTestBody, rawResponse?: core.RawResponse) {
         super({
             message: "PropertyBasedErrorTest",