+++ seed/ts-sdk/imdb/branded-string-aliases/src/api/errors/NotFoundError.ts
@@ -5,6 +5,8 @@ import * as errors from "../../errors/index.js";
 import type * as SeedApi from "../index.js";
 
 export class NotFoundError extends errors.SeedApiError {
+    public declare readonly body: SeedApi.MovieId;
+
     constructor(body: SeedApi.MovieId, rawResponse?: core.RawResponse) {
         super({
             message: "NotFoundError",