+++ seed/ts-sdk/server-sent-event-examples/src/api/resources/completions/errors/BadRequestError.ts
@@ -4,6 +4,8 @@ import type * as core from "../../../../core/index.js";
 import * as errors from "../../../../errors/index.js";
 
 export class BadRequestError extends errors.SeedServerSentEventsError {
+    public declare readonly body: string;
+
     constructor(body: string, rawResponse?: core.RawResponse) {
         super({
             message: "BadRequestError",